Refactor `ResNetBackbone` for readability (#1839)
* Self-explanatory

* Update resnet_v1_backbone.py

---------
